requirements.txt
anyio==3.3.1
attrs==19.3.0
Automat==0.8.0
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==2.0.4
Click==7.0
cloud-init==21.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
coverage==5.5
cryptography==2.8
dbus-python==1.2.16
deprecation==2.1.0
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
eyed3==0.9.6
filetype==1.0.7
h11==0.12.0
httpcore==0.13.7
httplib2==0.14.0
httpx==0.19.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==1.1.0
more-itertools==4.2.0
netifaces==0.10.4
oauthlib==3.1.0
packaging==21.0
pexpect==4.6.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pydub==0.25.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.6
python-debian===0.1.36ubuntu1
python-dotenv==0.19.0
pytube==11.0.1
PyYAML==5.3.1
requests==2.26.0
requests-unixsocket==0.2.0
rfc3986==1.5.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.16.0
sniffio==1.2.0
sos==4.1
spotipy==2.19.0
ssh-import-id==5.10
systemd-python==234
toml==0.10.2
Twisted==18.9.0
ubuntu-advantage-tools==27.2
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.6
wadllib==1.3.3
youtube-search-python==1.4.7
zipp==1.0.0
zope.interface==4.7.1
